+++ SurfaceTopography/IO/MI.py
@@ -27,11 +27,12 @@
 
 import numpy as np
 
+from ..Exceptions import CorruptFile, MetadataAlreadyFixedByFile
 from ..UniformLineScanAndTopography import Topography
 from ..UnitConversion import mangle_length_unit_utf8
 
 from .common import OpenFromAny
-from .Reader import ReaderBase, CorruptFile, ChannelInfo, MetadataAlreadyFixedByFile
+from .Reader import ReaderBase, ChannelInfo
 
 image_head = b'fileType      Image\n'
 spec_head = b'fileType      Spectroscopy\n'