+++ SurfaceTopography/IO/Text.py
@@ -29,13 +29,13 @@ import re
 
 import numpy as np
 
+from ..Exceptions import MetadataAlreadyFixedByFile
 from ..HeightContainer import UniformTopographyInterface
 from ..NonuniformLineScan import NonuniformLineScan
 from ..UniformLineScanAndTopography import Topography, UniformLineScan
 from ..UnitConversion import length_units, mangle_length_unit_utf8
 from .common import CHANNEL_NAME_INFO_KEY, text
 from .FromFile import make_wrapped_reader
-from .Reader import MetadataAlreadyFixedByFile
 
 ###