+++ SurfaceTopography/IO/__init__.py
@@ -27,6 +27,9 @@
 
 import os
 
+from ..Exceptions import CannotDetectFileFormat, CorruptFile, MetadataAlreadyFixedByFile, ReadFileError, \
+    UnknownFileFormatGiven  # noqa: F401
+
 # Old-style readers
 from .FromFile import HGTReader, OPDReader, X3PReader
 from .Text import AscReader, XYZReader
@@ -45,8 +48,7 @@ from .ZON import ZONReader
 # Only writers
 import SurfaceTopography.IO.DZI  # noqa: F401
 
-from .Reader import UnknownFileFormatGiven, CannotDetectFileFormat, MetadataAlreadyFixedByFile,\
-    FileFormatMismatch, CorruptFile, ReaderBase  # noqa: F401
+from .Reader import ReaderBase  # noqa: F401
 
 readers = [
     # XYZ must come before ASC, because 2D XYZ is a specialized ASC