+++ SurfaceTopography/Support/Regression.py
@@ -33,6 +33,8 @@ import logging
 
 import numpy as np
 
+from ..Exceptions import NoReliableDataError
+
 _log = logging.Logger(__name__)
 
 
@@ -297,6 +299,9 @@ def resample(x, values, collocation='log', nb_points=None, min_value=None, max_v
         max_value = np.max(x)
     if min_value is None:
         min_value = np.min(x)
+
+    if min_value >= max_value:
+        raise NoReliableDataError('Data contains no reliable data.')
 
     collocation_points, bin_edges = make_grid(collocation, min_value, max_value, nb_points=nb_points,
                                               nb_points_per_decade=nb_points_per_decade)